Use a specific card for integration test
Apparently the transactions go through with all the cards in the test
document but reports only report on transactions made with a specific
test card. Also there is about a 30-40 second delay between transaction
hitting the OLTP system to when it is available for reporting. So we
sleep for 42 seconds during the integration tests.